optimize: Encapsulate RightMenu Component And Refactor Rendering Method
